+++ DirectX12CG/Engin/Pipeline/RootParameter.h
@@ -12,9 +12,9 @@ namespace MCB
 		~RootParameter();
 
 		void SetRootParam(D3D12_ROOT_PARAMETER_TYPE paramType, int ShaderRegister, int RegisterSpace,
-			D3D12_SHADER_VISIBILITY shaderVisibility, int NumDescriptorRanges, unsigned int descriptorIndex = 0);
+			D3D12_SHADER_VISIBILITY shaderVisibility, int NumDescriptorRanges, size_t descriptorIndex = 0);
 
-		void SetRootParam(D3D12_ROOT_PARAMETER_TYPE paramType, int ShaderRegister = 0, unsigned int descriptorIndex = 0);
+		void SetRootParam(D3D12_ROOT_PARAMETER_TYPE paramType, int ShaderRegister = 0, size_t descriptorIndex = 0);
 
 	};
